+++ services/search-builder.js
@@ -5,33 +5,45 @@ var utils = require('../utils/schema');
 
 function SearchBuilder(model, opts, params, searchFields) {
   var schema = Interface.Schemas.schemas[utils.getModelName(model)];
+  var fieldsSearched = [];
+
+  this.getFieldsSearched = function () {
+    return fieldsSearched;
+  };
 
   this.getConditions = function () {
+
+    var orQuery = { $or: [] };
+
+    function pushCondition(condition, fieldName) {
+      orQuery.$or.push(condition);
+      fieldsSearched.push(fieldName);
+    }
+
     if (new RegExp('^[0-9a-fA-F]{24}$').test(params.search)) {
       return { _id: params.search };
     } else {
-      var orQuery = { $or: [] };
 
       _.each(model.schema.paths, function (value, key) {
         if (searchFields && searchFields.indexOf(value.path) === -1) {
           return null;
         }
-        var q = {};
+        var condition = {};
 
         if (value.instance === 'ObjectID') {
           try {
-            q[key] = opts.mongoose.Types.ObjectId(params.search);
-            orQuery.$or.push(q);
+            condition[key] = opts.mongoose.Types.ObjectId(params.search);
+            pushCondition(condition, key);
           } catch(error) { return null; }
         } else if (value.instance === 'String') {
-          q[key] = new RegExp('.*' + params.search + '.*', 'i');
-          orQuery.$or.push(q);
+          condition[key] = new RegExp('.*' + params.search + '.*', 'i');
+          pushCondition(condition, key);
         } else if (value.instance === 'Array') {
           var field = _.findWhere(schema.fields, { field: key });
           if (field && _.isArray(field.type) && field.type[0] === 'String' &&
             !field.reference) {
-            q[key] = new RegExp('.*' + params.search + '.*', 'i');
-            orQuery.$or.push(q);
+            condition[key] = new RegExp('.*' + params.search + '.*', 'i');
+            pushCondition(condition, key);
           } else if (field && _.isArray(field.type) &&
             !field.reference && parseInt(params.searchExtended)) {
             var elemMatch = { $elemMatch: { $or: [], } };
@@ -49,13 +61,13 @@ function SearchBuilder(model, opts, params, searchFields) {
                 elemMatch.$elemMatch.$or.push(query);
               }
             });
-            q[key] = elemMatch;
-            orQuery.$or.push(q);
+            condition[key] = elemMatch;
+            pushCondition(condition, key);
           }
         }
       });
 
-      return orQuery.$or.length ? orQuery : { _id: null };
+      return orQuery.$or.length ? orQuery : {};
     }
   };