+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Search - Add tests for search on selected field.
+- Search - Display highlighted matches on table view when searching.
 
 ### Changed
 - Technical - Switch the linter to eslint.
@@ -10,6 +11,9 @@
 
 ### Fixed
 - Technical - Use local packages for npm scripts.
+
+### Fixed
+- Search - Return empty result instead of the whole list of records if no field can be searched on the collection.
 
 ## RELEASE 2.8.4 - 2018-04-30
 ### Fixed